backend/requirements.txt
Django==5.1
djangorestframework==3.20.4
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
django-environ==0.11.0
python-dotenv==1.0.0
django-cors-headers==4.0.0
django-tailwind==3.0.0
pytz
